feat(game-loop): D1 — widen Lab header seed to u64, extract header_tag_labels
ShellSession::seed / LabInput::seed / draw_header's seed parameter move
i32 -> u64 end-to-end (issue #43 Scope 14), so a master seed exceeding
i32::MAX round-trips instead of truncating. Extracts the inline
format!("seed {seed}") into a pure header_tag_labels(seed: u64) -> Vec<String>
that draw_header calls, one Tag per entry — mirrors the oracle_tile_strings
precedent so the label is assertable without an egui::Context/Harness (AC15).

gp-game's session-seed cast (i32 truncating cast + fallback) is now a plain
Option<u64>::unwrap_or. Rendered seeds (42, 7) format identically as u64, so
no golden moves.

+++ crates/render/src/app.rs
@@ -408,7 +408,7 @@ pub struct ShellSession<'a> {
     /// The header validity flag — `Lab`.
     pub valid: bool,
     /// The header `seed <N>` tag value — `Lab`.
-    pub seed: i32,
+    pub seed: u64,
     /// The rank-ordered standings slice — `Results`.
     pub standings: &'a [StandingEntry],
     /// The race summary metrics — `Results`.
